chapter_summarization/requirements.txt
pika
sumy
numpy
psycopg2
pdfplumber
python-dotenv
google-cloud-storage